requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
attrdict=2.0.1=pypi_0
gnutls=3.6.15=he1e5248_0
gpustat=1.1.1=pypi_0
gputil=1.4.0=pypi_0
humanize=4.9.0=pypi_0
numba=0.59.0rc1=pypi_0
numpy=1.26.3=pypi_0
opencv-python=4.9.0.80=pypi_0
pesq=0.0.4=pypi_0
pillow=10.2.0=pypi_0
pip=24.0=py39h06a4308_0
pseq=2.1.5=pypi_0
psutil=5.9.8=pypi_0
pyroomacoustics=0.7.3=pypi_0
python=3.9.18=h955ad1f_0
pytorch=2.1.2=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
torchaudio=2.1.2=py39_cu118
torchvision=0.16.2=py39_cu118
tqdm=4.66.1=pypi_0
wheel=0.43.0=py39h06a4308_0
yaml=0.2.5=h7b6447c_0
